Move binary_search test utilities to support folder.
